backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpgsounddesk-backend"
version = "0.1.0"
description = "RPGSoundDesk backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "redis>=5.0.0,<6.0.0",
  "boto3>=1.34.0,<2.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "passlib>=1.7.4,<2.0.0",
  "email-validator>=2.2.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
